+++ src/assets/microgrid.jl
@@ -14,7 +14,7 @@ sizing and simulating.
 * *optionnal* : the minimum renewable_share [default = 0.]"""
 mutable struct GlobalParameters
     ns::Int64 # number of [nh, ny] scenarios
-    Δh::Int64 # operations time step in hours
+    Δh::Float64 # operations time step in hours
     nh::Int64 # number of operation stages
     Δy::Int64 # investment time step in years
     ny::Int64 # number of investment stages
@@ -22,7 +22,7 @@ mutable struct GlobalParameters
     renewable_share::Float64 # share of renewables [0,1]
 
     GlobalParameters(nh, ny, ns;
-                Δh = 1,
+                Δh = 1.,
                 Δy = 1,
                 τ = 0.045,
                 renewable_share = 0.) = renewable_share == 1. ? new(ns, Δh, nh, Δy, ny, τ, 0.9999) : new(ns, Δh, nh, Δy, ny, τ, renewable_share)
